@@ -20,7 +20,7 @@
     }
   ],
   "require": {
-    "php": ">=7.4",
+    "php": "^8.1",
     "ext-SimpleXML": "*",
     "ext-dom": "*",
     "ext-iconv": "*",
@@ -29,32 +29,16 @@
     "ext-mbstring": "*",
     "lib-iconv": "*",
     "lib-libxml": "*",
-    "doctrine/dbal": "~2.3",
+    "doctrine/dbal": "^3.6",
     "phpoption/phpoption": "*",
     "vlucas/phpdotenv": "^5.2",
     "guzzlehttp/guzzle": "^7.0.1",
     "composer/composer": "*",
-    "illuminate/cache": "8.*",
-    "illuminate/redis": "8.*",
-    "illuminate/config": "8.*",
-    "illuminate/console": "8.*",
-    "illuminate/container": "8.*",
-    "illuminate/database": "8.*",
-    "illuminate/events": "8.*",
-    "illuminate/filesystem": "8.*",
-    "illuminate/log": "8.*",
-    "illuminate/pagination": "8.*",
-    "illuminate/routing": "8.*",
-    "illuminate/support": "8.*",
-    "illuminate/translation": "8.*",
-    "illuminate/view": "8.*",
-    "illuminate/validation": "8.*",
-    "illuminate/http": "8.*",
-    "league/flysystem": "^1.0.8",
+    "laravel/framework": "9.52.*",
     "phpmailer/phpmailer": "6.*",
     "tracy/tracy": "2.9.*",
     "wikimedia/composer-merge-plugin": "2.*",
-    "symfony/process": "^5.1",
+    "symfony/process": "^6.0",
     "dmitry-suffi/redis-session-handler": "*",
     "rosell-dk/webp-convert": "*",
     "predis/predis": "*",
@@ -62,7 +46,7 @@
     "evocms-services/user-manager": "1.*",
     "evocms-services/document-manager": "1.*",
     "evolution-cms/salo": "1.*",
-    "simplepie/simplepie": "1.*",
+    "simplepie/simplepie": "^1.8",
     "pathologic/modxapi": "*"
   },
   "require-dev": {
@@ -109,7 +93,6 @@
       "functions/actions/mutate_plugin.php",
       "functions/actions/settings.php",
       "functions/helper.php",
-      "functions/laravel.php",
       "functions/nodes.php",
       "functions/preload.php",
       "functions/processors.php",
