@@ -18,7 +18,6 @@
         "queryobject",
         "sasql",
         "sql",
-        "sqlanywhere",
         "sqlite",
         "sqlserver",
         "sqlsrv"
@@ -32,22 +31,26 @@
         {"name": "Jonathan Wage", "email": "jonwage@gmail.com"}
     ],
     "require": {
-        "php": "^7.1 || ^8",
-        "ext-pdo": "*",
-        "doctrine/cache": "^1.0|^2.0",
+        "php": "^7.4 || ^8.0",
+        "composer-runtime-api": "^2",
+        "doctrine/cache": "^1.11|^2.0",
         "doctrine/deprecations": "^0.5.3|^1",
-        "doctrine/event-manager": "^1.0"
+        "doctrine/event-manager": "^1|^2",
+        "psr/cache": "^1|^2|^3",
+        "psr/log": "^1|^2|^3"
     },
     "require-dev": {
-        "doctrine/coding-standard": "9.0.0",
-        "jetbrains/phpstorm-stubs": "2021.1",
-        "phpstan/phpstan": "1.4.6",
-        "phpunit/phpunit": "^7.5.20|^8.5|9.5.16",
-        "psalm/plugin-phpunit": "0.16.1",
-        "squizlabs/php_codesniffer": "3.6.2",
-        "symfony/cache": "^4.4",
-        "symfony/console": "^2.0.5|^3.0|^4.0|^5.0",
-        "vimeo/psalm": "4.22.0"
+        "doctrine/coding-standard": "11.1.0",
+        "fig/log-test": "^1",
+        "jetbrains/phpstorm-stubs": "2022.3",
+        "phpstan/phpstan": "1.10.9",
+        "phpstan/phpstan-strict-rules": "^1.5",
+        "phpunit/phpunit": "9.6.6",
+        "psalm/plugin-phpunit": "0.18.4",
+        "squizlabs/php_codesniffer": "3.7.2",
+        "symfony/cache": "^5.4|^6.0",
+        "symfony/console": "^4.4|^5.4|^6.0",
+        "vimeo/psalm": "4.30.0"
     },
     "suggest": {
         "symfony/console": "For helpful console commands such as SQL execution and import of files."
@@ -61,9 +64,9 @@
         }
     },
     "autoload": {
-        "psr-4": { "Doctrine\\DBAL\\": "lib/Doctrine/DBAL" }
+        "psr-4": { "Doctrine\\DBAL\\": "src" }
     },
     "autoload-dev": {
-        "psr-4": { "Doctrine\\Tests\\": "tests/Doctrine/Tests" }
+        "psr-4": { "Doctrine\\DBAL\\Tests\\": "tests" }
     }
 }
