@@ -38,14 +38,13 @@
     ],
     "homepage": "https://www.doctrine-project.org/projects/event-manager.html",
     "require": {
-        "php": "^7.1 || ^8.0",
-        "doctrine/deprecations": "^0.5.3 || ^1"
+        "php": "^8.1"
     },
     "require-dev": {
-        "doctrine/coding-standard": "^9 || ^10",
-        "phpstan/phpstan": "~1.4.10 || ^1.8.8",
-        "phpunit/phpunit": "^7.5 || ^8.5 || ^9.5",
-        "vimeo/psalm": "^4.24"
+        "doctrine/coding-standard": "^10",
+        "phpstan/phpstan": "^1.8.8",
+        "phpunit/phpunit": "^9.5",
+        "vimeo/psalm": "^4.28"
     },
     "conflict": {
         "doctrine/common": "<2.9"
