@@ -1,68 +1,67 @@
 {
     "name": "league/flysystem",
-    "type": "library",
-    "description": "Filesystem abstraction: Many filesystems, one API.",
+    "description": "File storage abstraction for PHP",
     "keywords": [
-        "filesystem", "filesystems", "files", "storage", "dropbox", "aws",
-        "abstraction", "s3", "ftp", "sftp", "remote", "webdav",
-        "file systems", "cloud", "cloud files", "rackspace", "copy.com"
+        "filesystem", "filesystems", "files", "storage", "aws",
+        "s3", "ftp", "sftp", "webdav", "file", "cloud"
     ],
-    "funding": [
-        {
-            "type": "other",
-            "url": "https://offset.earth/frankdejonge"
+    "scripts": {
+        "phpstan": "vendor/bin/phpstan analyse -l 6 src"
+    },
+    "type": "library",
+    "minimum-stability": "dev",
+    "prefer-stable": true,
+    "autoload": {
+        "psr-4": {
+            "League\\Flysystem\\": "src"
         }
-    ],
+    },
+    "require": {
+        "php": "^8.0.2",
+        "league/flysystem-local":  "^3.0.0",
+        "league/mime-type-detection": "^1.0.0"
+    },
+    "require-dev": {
+        "ext-zip": "*",
+        "ext-fileinfo": "*",
+        "ext-ftp": "*",
+        "microsoft/azure-storage-blob": "^1.1",
+        "phpunit/phpunit": "^9.5.11",
+        "phpstan/phpstan": "^0.12.26",
+        "phpseclib/phpseclib": "^3.0.14",
+        "aws/aws-sdk-php": "^3.220.0",
+        "composer/semver": "^3.0",
+        "friendsofphp/php-cs-fixer": "^3.5",
+        "google/cloud-storage": "^1.23",
+        "async-aws/s3": "^1.5",
+        "async-aws/simple-s3": "^1.1",
+        "sabre/dav": "^4.3.1"
+    },
+    "conflict": {
+        "symfony/http-client": "<5.2",
+        "guzzlehttp/ringphp": "<1.1.1",
+        "guzzlehttp/guzzle": "<7.0",
+        "aws/aws-sdk-php": "3.209.31 || 3.210.0",
+        "phpseclib/phpseclib": "3.0.15"
+    },
     "license": "MIT",
     "authors": [
         {
             "name": "Frank de Jonge",
-            "email": "info@frenky.net"
+            "email": "info@frankdejonge.nl"
         }
     ],
-    "require": {
-        "php": "^7.2.5 || ^8.0",
-        "ext-fileinfo": "*",
-        "league/mime-type-detection": "^1.3"
-    },
-    "require-dev": {
-        "phpspec/prophecy": "^1.11.1",
-        "phpunit/phpunit": "^8.5.8"
-    },
-    "autoload": {
-        "psr-4": {
-            "League\\Flysystem\\": "src/"
+    "repositories": [
+        {
+            "type": "package",
+            "package": {
+                "name": "league/flysystem-local",
+                "version": "3.0.0",
+                "dist": {
+                    "type": "path",
+                    "url": "src/Local"
+                }
+            }
         }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "League\\Flysystem\\Stub\\": "stub/"
-        }
-    },
-    "suggest": {
-        "league/flysystem-eventable-filesystem": "Allows you to use EventableFilesystem",
-        "league/flysystem-rackspace": "Allows you to use Rackspace Cloud Files",
-        "league/flysystem-azure": "Allows you to use Windows Azure Blob storage",
-        "league/flysystem-webdav": "Allows you to use WebDAV storage",
-        "league/flysystem-aws-s3-v2": "Allows you to use S3 storage with AWS SDK v2",
-        "league/flysystem-aws-s3-v3": "Allows you to use S3 storage with AWS SDK v3",
-        "spatie/flysystem-dropbox": "Allows you to use Dropbox storage",
-        "srmklive/flysystem-dropbox-v2": "Allows you to use Dropbox storage for PHP 5 applications",
-        "league/flysystem-cached-adapter": "Flysystem adapter decorator for metadata caching",
-        "ext-ftp": "Allows you to use FTP server storage",
-        "ext-openssl": "Allows you to use FTPS server storage",
-        "league/flysystem-sftp": "Allows you to use SFTP server storage via phpseclib",
-        "league/flysystem-ziparchive": "Allows you to use ZipArchive adapter"
-    },
-    "conflict": {
-        "league/flysystem-sftp": "<1.0.6"
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "1.1-dev"
-        }
-    },
-    "scripts": {
-        "phpstan": "php phpstan.php"
-    }
+    ]
 }
