@@ -16,25 +16,23 @@
         }
     ],
     "require": {
-        "php": ">=7.2.5",
+        "php": ">=8.1",
         "symfony/deprecation-contracts": "^2.1|^3",
         "symfony/polyfill-mbstring": "~1.0",
-        "symfony/polyfill-php73": "^1.9",
-        "symfony/polyfill-php80": "^1.16",
         "symfony/service-contracts": "^1.1|^2|^3",
-        "symfony/string": "^5.1|^6.0"
+        "symfony/string": "^5.4|^6.0"
     },
     "require-dev": {
-        "symfony/config": "^4.4|^5.0|^6.0",
-        "symfony/event-dispatcher": "^4.4|^5.0|^6.0",
-        "symfony/dependency-injection": "^4.4|^5.0|^6.0",
-        "symfony/lock": "^4.4|^5.0|^6.0",
-        "symfony/process": "^4.4|^5.0|^6.0",
-        "symfony/var-dumper": "^4.4|^5.0|^6.0",
-        "psr/log": "^1|^2"
+        "symfony/config": "^5.4|^6.0",
+        "symfony/event-dispatcher": "^5.4|^6.0",
+        "symfony/dependency-injection": "^5.4|^6.0",
+        "symfony/lock": "^5.4|^6.0",
+        "symfony/process": "^5.4|^6.0",
+        "symfony/var-dumper": "^5.4|^6.0",
+        "psr/log": "^1|^2|^3"
     },
     "provide": {
-        "psr/log-implementation": "1.0|2.0"
+        "psr/log-implementation": "1.0|2.0|3.0"
     },
     "suggest": {
         "symfony/event-dispatcher": "",
@@ -43,12 +41,11 @@
         "psr/log": "For using the console logger"
     },
     "conflict": {
-        "psr/log": ">=3",
-        "symfony/dependency-injection": "<4.4",
-        "symfony/dotenv": "<5.1",
-        "symfony/event-dispatcher": "<4.4",
-        "symfony/lock": "<4.4",
-        "symfony/process": "<4.4"
+        "symfony/dependency-injection": "<5.4",
+        "symfony/dotenv": "<5.4",
+        "symfony/event-dispatcher": "<5.4",
+        "symfony/lock": "<5.4",
+        "symfony/process": "<5.4"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Console\\": "" },
