@@ -15,7 +15,7 @@
         }
     ],
     "require": {
-        "php": ">=7.1"
+        "php": ">=8.1"
     },
     "autoload": {
         "files": [
@@ -25,7 +25,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-main": "2.5-dev"
+            "dev-main": "3.3-dev"
         },
         "thanks": {
             "name": "symfony/contracts",
