@@ -1,23 +1,24 @@
 <?php namespace EvolutionCMS\Controllers;
 
-use EvolutionCMS\Interfaces\ManagerTheme;
-use EvolutionCMS\Models;
+use EvolutionCMS\Facades\ManagerTheme;
+use EvolutionCMS\Interfaces\ManagerTheme\PageControllerInterface;
+use EvolutionCMS\Models\EventLog;
 
-class EventLogDetails extends AbstractController implements ManagerTheme\PageControllerInterface
+class EventLogDetails extends AbstractController implements PageControllerInterface
 {
-    protected $view = 'page.eventlog_details';
+    protected string $view = 'page.eventlog_details';
 
     /**
      * {@inheritdoc}
      */
     public function canView(): bool
     {
-        return $this->managerTheme->getCore()->hasPermission('view_eventlog');
+        return ManagerTheme::getCore()->hasPermission('view_eventlog');
     }
 
     public function process() : bool
     {
-        $this->parameters['log'] = Models\EventLog::findOrNew($this->getElementId());
+        $this->parameters['log'] = EventLog::findOrNew($this->getElementId());
 
         return true;
     }
