@@ -1,12 +1,13 @@
 <?php namespace EvolutionCMS\Controllers;
 
-use EvolutionCMS\Interfaces\ManagerTheme;
+use EvolutionCMS\Facades\ManagerTheme;
+use EvolutionCMS\Interfaces\ManagerTheme\PageControllerInterface;
 use EvolutionCMS\Interfaces\ManagerThemeInterface;
 use EvolutionCMS\Models\SiteModule;
 
-class Frame extends AbstractController implements ManagerTheme\PageControllerInterface
+class Frame extends AbstractController implements PageControllerInterface
 {
-    protected $view;
+    protected string $view;
 
     protected $frame;
 
@@ -20,7 +21,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
         if ($this->frame > 9) {
             // this is to stop the debug thingy being attached to the framesets
-            $this->managerTheme->getCore()->setConfig('enable_debug', false);
+            ManagerTheme::getCore()->setConfig('enable_debug', false);
         }
         if (!empty($this->frame)) {
             $this->setView('frame.' . $this->frame);
@@ -45,15 +46,15 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->initSession();
 
         // invoke OnManagerPreFrameLoader
-        $this->managerTheme->getCore()->invokeEvent(
+        ManagerTheme::getCore()->invokeEvent(
             'OnManagerPreFrameLoader',
             [
-                'action' => $this->managerTheme->getActionId(),
+                'action' => ManagerTheme::getActionId(),
             ]
         );
 
         $body_class = '';
-        $tree_width = $this->managerTheme->getCore()->getConfig('manager_tree_width');
+        $tree_width = ManagerTheme::getCore()->getConfig('manager_tree_width');
         $this->parameters['tree_min_width'] = 0;
 
         if (isset($_COOKIE['MODX_widthSideBar'])) {
@@ -70,60 +71,60 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $theme_modes = ['', 'lightness', 'light', 'dark', 'darkness'];
         if (isset($_COOKIE['MODX_themeMode']) && !empty($theme_modes[$_COOKIE['MODX_themeMode']])) {
             $body_class .= ' ' . $theme_modes[$_COOKIE['MODX_themeMode']];
-        } elseif (!empty($theme_modes[$this->managerTheme->getCore()->getConfig('manager_theme_mode')])) {
-            $body_class .= ' ' . $theme_modes[$this->managerTheme->getCore()->getConfig('manager_theme_mode')];
+        } elseif (!empty($theme_modes[ManagerTheme::getCore()->getConfig('manager_theme_mode')])) {
+            $body_class .= ' ' . $theme_modes[ManagerTheme::getCore()->getConfig('manager_theme_mode')];
         }
 
-        $navbar_position = $this->managerTheme->getCore()->getConfig('manager_menu_position');
+        $navbar_position = ManagerTheme::getCore()->getConfig('manager_menu_position');
         if ($navbar_position === 'left') {
             $body_class .= ' navbar-left navbar-left-icon-and-text';
         }
 
-        if (isset($this->managerTheme->getCore()->pluginCache['ElementsInTree'])) {
+        if (isset(ManagerTheme::getCore()->pluginCache['ElementsInTree'])) {
             $body_class .= ' ElementsInTree';
         }
 
         $this->parameters['body_class'] = $body_class;
 
         $unlockTranslations = [
-            'msg' => $this->managerTheme->getLexicon('unlock_element_id_warning'),
-            'type1' => $this->managerTheme->getLexicon('lock_element_type_1'),
-            'type2' => $this->managerTheme->getLexicon('lock_element_type_2'),
-            'type3' => $this->managerTheme->getLexicon('lock_element_type_3'),
-            'type4' => $this->managerTheme->getLexicon('lock_element_type_4'),
-            'type5' => $this->managerTheme->getLexicon('lock_element_type_5'),
-            'type6' => $this->managerTheme->getLexicon('lock_element_type_6'),
-            'type7' => $this->managerTheme->getLexicon('lock_element_type_7'),
-            'type8' => $this->managerTheme->getLexicon('lock_element_type_8'),
+            'msg' => ManagerTheme::getLexicon('unlock_element_id_warning'),
+            'type1' => ManagerTheme::getLexicon('lock_element_type_1'),
+            'type2' => ManagerTheme::getLexicon('lock_element_type_2'),
+            'type3' => ManagerTheme::getLexicon('lock_element_type_3'),
+            'type4' => ManagerTheme::getLexicon('lock_element_type_4'),
+            'type5' => ManagerTheme::getLexicon('lock_element_type_5'),
+            'type6' => ManagerTheme::getLexicon('lock_element_type_6'),
+            'type7' => ManagerTheme::getLexicon('lock_element_type_7'),
+            'type8' => ManagerTheme::getLexicon('lock_element_type_8'),
         ];
 
         foreach ($unlockTranslations as $key => $value) {
             $unlockTranslations[$key] = iconv(
-                $this->managerTheme->getCore()->getConfig('modx_charset'),
+                ManagerTheme::getCore()->getConfig('modx_charset'),
                 'utf-8',
                 $value
             );
         }
         $this->parameters['unlockTranslations'] = $unlockTranslations;
 
-        $user = $this->managerTheme->getCore()->getUserInfo($this->managerTheme->getCore()->getLoginUserID('mgr'));
+        $user = ManagerTheme::getCore()->getUserInfo(ManagerTheme::getCore()->getLoginUserID('mgr'));
         if ((isset($user['which_browser']) && $user['which_browser'] == 'default') || (!isset($user['which_browser']))) {
-            $user['which_browser'] = $this->managerTheme->getCore()->getConfig('which_browser');
+            $user['which_browser'] = ManagerTheme::getCore()->getConfig('which_browser');
         }
         $this->parameters['user'] = $user;
 
         $this->registerCss();
 
-        $flag = $user['role'] == 1 || $this->managerTheme->getCore()->hasAnyPermissions([
+        $flag = $user['role'] == 1 || ManagerTheme::getCore()->hasAnyPermissions([
             'edit_template',
             'edit_chunk',
             'edit_snippet',
             'edit_plugin',
         ]);
 
-        $this->managerTheme->getCore()->setConfig(
+        ManagerTheme::getCore()->setConfig(
             'global_tabs',
-            (int) ($this->managerTheme->getCore()->getConfig('global_tabs') && $flag)
+            (int) (ManagerTheme::getCore()->getConfig('global_tabs') && $flag)
         );
 
         $this->makeMenu();
@@ -158,9 +159,9 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function registerCss(): string
     {
-        $this->parameters['css'] = $this->managerTheme->getThemeDir(false) . 'css/page.css?v=' . EVO_INSTALL_TIME;
+        $this->parameters['css'] = ManagerTheme::getThemeDir(false) . 'css/page.css?v=' . EVO_INSTALL_TIME;
 
-        $themeDir = $this->managerTheme->getThemeDir();
+        $themeDir = ManagerTheme::getThemeDir();
 
         if (file_exists($themeDir . 'CSSMinify.php')) {
             if (!file_exists($themeDir . 'css/styles.min.css') && is_writable($themeDir . 'css')) {
@@ -175,7 +176,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             }
 
             if (file_exists($themeDir . 'css/styles.min.css')) {
-                $this->parameters['css'] = $this->managerTheme->getThemeDir(false) . 'css/styles.min.css?v=' . EVO_INSTALL_TIME;
+                $this->parameters['css'] = ManagerTheme::getThemeDir(false) . 'css/styles.min.css?v=' . EVO_INSTALL_TIME;
             }
         }
 
@@ -204,7 +205,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             ->menuRemoveLocks()
             ->menuUpdateTree();
 
-        $menu = $this->managerTheme->getCore()->invokeEvent('OnManagerMenuPrerender', ['menu' => $this->sitemenu]);
+        $menu = ManagerTheme::getCore()->invokeEvent('OnManagerMenuPrerender', ['menu' => $this->sitemenu]);
         if (\is_array($menu)) {
             $newmenu = [];
             foreach ($menu as $item) {
@@ -240,9 +241,9 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['bars'] = [
             'bars',
             'main',
-            '<i class="' . $this->managerTheme->getStyle('icon_bars') . '"></i>',
+            '<i class="' . ManagerTheme::getStyle('icon_bars') . '"></i>',
             'javascript:;',
-            $this->managerTheme->getLexicon('home'),
+            ManagerTheme::getLexicon('home'),
             'modx.resizer.toggle(); return false;',
             ' return false;',
             '',
@@ -259,9 +260,9 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['site'] = [
             'site',
             'main',
-            '<i class="' . $this->managerTheme->getStyle('icon_tachometer') . '"></i><span class="menu-item-text">' . $this->managerTheme->getLexicon('home') . '</span>',
+            '<i class="' . ManagerTheme::getStyle('icon_tachometer') . '"></i><span class="menu-item-text">' . ManagerTheme::getLexicon('home') . '</span>',
             'index.php?a=2',
-            $this->managerTheme->getLexicon('home'),
+            ManagerTheme::getLexicon('home'),
             '',
             '',
             'main',
@@ -275,7 +276,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuElementTypes()
     {
-        $flag = $this->managerTheme->getCore()->hasAnyPermissions(
+        $flag = ManagerTheme::getCore()->hasAnyPermissions(
             ['edit_template', 'edit_snippet', 'edit_chunk', 'edit_plugin', 'category_manager', 'file_manager']
         );
 
@@ -286,9 +287,9 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['elements'] = [
             'elements',
             'main',
-            '<i class="' . $this->managerTheme->getStyle('icon_elements') . '"></i><span class="menu-item-text">' . $this->managerTheme->getLexicon('elements') . '</span>',
+            '<i class="' . ManagerTheme::getStyle('icon_elements') . '"></i><span class="menu-item-text">' . ManagerTheme::getLexicon('elements') . '</span>',
             'javascript:;',
-            $this->managerTheme->getLexicon('elements'),
+            ManagerTheme::getLexicon('elements'),
             ' return false;',
             '',
             '',
@@ -302,16 +303,16 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuModules()
     {
-        if (!$this->managerTheme->getCore()->hasPermission('exec_module')) {
+        if (!ManagerTheme::getCore()->hasPermission('exec_module')) {
             return $this;
         }
 
         $this->sitemenu['modules'] = [
             'modules',
             'main',
-            '<i class="' . $this->managerTheme->getStyle('icon_modules') . '"></i><span class="menu-item-text">' . $this->managerTheme->getLexicon('modules') . '</span>',
+            '<i class="' . ManagerTheme::getStyle('icon_modules') . '"></i><span class="menu-item-text">' . ManagerTheme::getLexicon('modules') . '</span>',
             'javascript:;',
-            $this->managerTheme->getLexicon('modules'),
+            ManagerTheme::getLexicon('modules'),
             ' return false;',
             '',
             '',
@@ -325,7 +326,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuUsers()
     {
-        $flag = $this->managerTheme->getCore()->hasAnyPermissions(
+        $flag = ManagerTheme::getCore()->hasAnyPermissions(
             ['edit_user', 'edit_role', 'manage_groups']
         );
 
@@ -336,9 +337,9 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['users'] = [
             'users',
             'main',
-            '<i class="' . $this->managerTheme->getStyle('icon_users') . '"></i><span class="menu-item-text">' . $this->managerTheme->getLexicon('users') . '</span>',
+            '<i class="' . ManagerTheme::getStyle('icon_users') . '"></i><span class="menu-item-text">' . ManagerTheme::getLexicon('users') . '</span>',
             'javascript:;',
-            $this->managerTheme->getLexicon('users'),
+            ManagerTheme::getLexicon('users'),
             ' return false;',
             'edit_user',
             '',
@@ -352,7 +353,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuTools()
     {
-        $flag = $this->managerTheme->getCore()->hasAnyPermissions(
+        $flag = ManagerTheme::getCore()->hasAnyPermissions(
             ['empty_cache', 'bk_manager', 'remove_locks', 'import_static', 'export_static']
         );
 
@@ -363,9 +364,9 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['tools'] = [
             'tools',
             'main',
-            '<i class="' . $this->managerTheme->getStyle('icon_wrench') . '"></i><span class="menu-item-text">' . $this->managerTheme->getLexicon('tools') . '</span>',
+            '<i class="' . ManagerTheme::getStyle('icon_wrench') . '"></i><span class="menu-item-text">' . ManagerTheme::getLexicon('tools') . '</span>',
             'javascript:;',
-            $this->managerTheme->getLexicon('tools'),
+            ManagerTheme::getLexicon('tools'),
             ' return false;',
             '',
             '',
@@ -391,16 +392,16 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuElementTemplates()
     {
-        if (!$this->managerTheme->getCore()->hasPermission('edit_template')) {
+        if (!ManagerTheme::getCore()->hasPermission('edit_template')) {
             return;
         }
 
         $this->sitemenu['element_templates'] = [
             'element_templates',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_template') . '"></i>' . $this->managerTheme->getLexicon('templates') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            '<i class="' . ManagerTheme::getStyle('icon_template') . '"></i>' . ManagerTheme::getLexicon('templates') . '<i class="' . ManagerTheme::getStyle('icon_angle_right') . ' toggle"></i>',
             'index.php?a=76&tab=0',
-            $this->managerTheme->getLexicon('templates'),
+            ManagerTheme::getLexicon('templates'),
             '',
             'new_template,edit_template',
             'main',
@@ -412,16 +413,16 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuElementTv()
     {
-        if (!$this->managerTheme->getCore()->hasAnyPermissions(['edit_template', 'edit_role'])) {
+        if (!ManagerTheme::getCore()->hasAnyPermissions(['edit_template', 'edit_role'])) {
             return;
         }
 
         $this->sitemenu['element_tplvars'] = [
             'element_tplvars',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_tv') . '"></i>' . $this->managerTheme->getLexicon('tmplvars') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            '<i class="' . ManagerTheme::getStyle('icon_tv') . '"></i>' . ManagerTheme::getLexicon('tmplvars') . '<i class="' . ManagerTheme::getStyle('icon_angle_right') . ' toggle"></i>',
             'index.php?a=76&tab=1',
-            $this->managerTheme->getLexicon('tmplvars'),
+            ManagerTheme::getLexicon('tmplvars'),
             '',
             'new_template,edit_template',
             'main',
@@ -433,16 +434,16 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuElementChunks()
     {
-        if (!$this->managerTheme->getCore()->hasPermission('edit_chunk')) {
+        if (!ManagerTheme::getCore()->hasPermission('edit_chunk')) {
             return;
         }
 
         $this->sitemenu['element_htmlsnippets'] = [
             'element_htmlsnippets',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_chunk') . '"></i>' . $this->managerTheme->getLexicon('htmlsnippets') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            '<i class="' . ManagerTheme::getStyle('icon_chunk') . '"></i>' . ManagerTheme::getLexicon('htmlsnippets') . '<i class="' . ManagerTheme::getStyle('icon_angle_right') . ' toggle"></i>',
             'index.php?a=76&tab=2',
-            $this->managerTheme->getLexicon('htmlsnippets'),
+            ManagerTheme::getLexicon('htmlsnippets'),
             '',
             'new_chunk,edit_chunk',
             'main',
@@ -454,16 +455,16 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuElementSnippets()
     {
-        if (!$this->managerTheme->getCore()->hasPermission('edit_snippet')) {
+        if (!ManagerTheme::getCore()->hasPermission('edit_snippet')) {
             return;
         }
 
         $this->sitemenu['element_snippets'] = [
             'element_snippets',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_code') . '"></i>' . $this->managerTheme->getLexicon('snippets') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            '<i class="' . ManagerTheme::getStyle('icon_code') . '"></i>' . ManagerTheme::getLexicon('snippets') . '<i class="' . ManagerTheme::getStyle('icon_angle_right') . ' toggle"></i>',
             'index.php?a=76&tab=3',
-            $this->managerTheme->getLexicon('snippets'),
+            ManagerTheme::getLexicon('snippets'),
             '',
             'new_snippet,edit_snippet',
             'main',
@@ -475,16 +476,16 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuElementPlugins()
     {
-        if (!$this->managerTheme->getCore()->hasPermission('edit_plugin')) {
+        if (!ManagerTheme::getCore()->hasPermission('edit_plugin')) {
             return;
         }
 
         $this->sitemenu['element_plugins'] = [
             'element_plugins',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_plugin') . '"></i>' . $this->managerTheme->getLexicon('plugins') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            '<i class="' . ManagerTheme::getStyle('icon_plugin') . '"></i>' . ManagerTheme::getLexicon('plugins') . '<i class="' . ManagerTheme::getStyle('icon_angle_right') . ' toggle"></i>',
             'index.php?a=76&tab=4',
-            $this->managerTheme->getLexicon('plugins'),
+            ManagerTheme::getLexicon('plugins'),
             '',
             'new_plugin,edit_plugin',
             'main',
@@ -496,16 +497,16 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuElementModules()
     {
-        if (!$this->managerTheme->getCore()->hasPermission('edit_module')) {
+        if (!ManagerTheme::getCore()->hasPermission('edit_module')) {
             return;
         }
 
         $this->sitemenu['element_modules'] = [
             'element_modules',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_module') . '"></i>' . $this->managerTheme->getLexicon('modules') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            '<i class="' . ManagerTheme::getStyle('icon_module') . '"></i>' . ManagerTheme::getLexicon('modules') . '<i class="' . ManagerTheme::getStyle('icon_angle_right') . ' toggle"></i>',
             'index.php?a=76&tab=5',
-            $this->managerTheme->getLexicon('modules'),
+            ManagerTheme::getLexicon('modules'),
             '',
             'new_module,edit_module',
             'main',
@@ -517,16 +518,16 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuFiles()
     {
-        if (!$this->managerTheme->getCore()->hasPermission('file_manager')) {
+        if (!ManagerTheme::getCore()->hasPermission('file_manager')) {
             return $this;
         }
 
         $this->sitemenu['manage_files'] = [
             'manage_files',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_folder_open') . '"></i>' . $this->managerTheme->getLexicon('files'),
+            '<i class="' . ManagerTheme::getStyle('icon_folder_open') . '"></i>' . ManagerTheme::getLexicon('files'),
             'index.php?a=31',
-            $this->managerTheme->getLexicon('files'),
+            ManagerTheme::getLexicon('files'),
             '',
             'file_manager',
             'main',
@@ -540,16 +541,16 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuCategories()
     {
-        if (!$this->managerTheme->getCore()->hasPermission('category_manager')) {
+        if (!ManagerTheme::getCore()->hasPermission('category_manager')) {
             return $this;
         }
 
         $this->sitemenu['manage_categories'] = [
             'manage_categories',
             'elements',
-            '<i class="' . $this->managerTheme->getStyle('icon_category') . '"></i>' . $this->managerTheme->getLexicon('categories'),
+            '<i class="' . ManagerTheme::getStyle('icon_category') . '"></i>' . ManagerTheme::getLexicon('categories'),
             'index.php?a=120',
-            $this->managerTheme->getLexicon('categories'),
+            ManagerTheme::getLexicon('categories'),
             '',
             'category_manager',
             'main',
@@ -563,7 +564,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuNewModule()
     {
-        $flag = $this->managerTheme->getCore()->hasAnyPermissions(
+        $flag = ManagerTheme::getCore()->hasAnyPermissions(
             ['new_module', 'edit_module', 'save_module']
         );
 
@@ -576,11 +577,11 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuRunModules()
     {
-        if ($this->managerTheme->getCore()->hasPermission('exec_module')) {
+        if (ManagerTheme::getCore()->hasPermission('exec_module')) {
             $items = [];
 
             // 1. modules from DB
-            if ($_SESSION['mgrRole'] != 1 && $this->managerTheme->getCore()->getConfig('use_udperms') === true) {
+            if ($_SESSION['mgrRole'] != 1 && ManagerTheme::getCore()->getConfig('use_udperms') === true) {
                 $modules = SiteModule::select('site_modules.id', 'site_modules.name', 'site_modules.icon', 'member_groups.member')
                     ->withoutProtected()
                     ->lockedView()
@@ -599,7 +600,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
                     $items['module' . $row['id']] = [
                         'module' . $row['id'],
                         'modules',
-                        ($row['icon'] != '' ? '<i class="' . e($row['icon']) . '"></i>' : '<i class="' . $this->managerTheme->getStyle('icon_module') . '"></i>') . e($row['name']),
+                        ($row['icon'] != '' ? '<i class="' . e($row['icon']) . '"></i>' : '<i class="' . ManagerTheme::getStyle('icon_module') . '"></i>') . e($row['name']),
                         'index.php?a=112&id=' . $row['id'],
                         e($row['name']),
                         '',
@@ -613,7 +614,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
             }
 
             // 2. modules from files
-            foreach ($this->managerTheme->getCore()->modulesFromFile as $module) {
+            foreach (ManagerTheme::getCore()->modulesFromFile as $module) {
                 if (!empty($module['properties']['hidden'])) {
                     continue;
                 }
@@ -621,7 +622,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
                 $items['module' . $module['id']] = [
                     'module' . $module['id'],
                     'modules',
-                    ($module['icon'] != '' ? '<i class="' . $module['icon'] . '"></i>' : '<i class="' . $this->managerTheme->getStyle('icon_module') . '"></i>') . $module['name'],
+                    ($module['icon'] != '' ? '<i class="' . $module['icon'] . '"></i>' : '<i class="' . ManagerTheme::getStyle('icon_module') . '"></i>') . $module['name'],
                     !empty($module['properties']['routes']) ? 'modules/' . $module['id'] : 'index.php?a=112&id=' . $module['id'],
                     $module['name'],
                     '',
@@ -647,7 +648,7 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuUserManagement()
     {
-        $flag = $this->managerTheme->getCore()->hasPermission('edit_user');
+        $flag = ManagerTheme::getCore()->hasPermission('edit_user');
 
         if (!$flag) {
             return $this;
@@ -656,9 +657,9 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['web_user_management_title'] = [
             'web_user_management_title',
             'users',
-            '<i class="' . $this->managerTheme->getStyle('icon_web_user') . '"></i>' . $this->managerTheme->getLexicon('web_user_management_title') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+            '<i class="' . ManagerTheme::getStyle('icon_web_user') . '"></i>' . ManagerTheme::getLexicon('web_user_management_title') . '<i class="' . ManagerTheme::getStyle('icon_angle_right') . ' toggle"></i>',
             'index.php?a=99',
-            $this->managerTheme->getLexicon('web_user_management_title'),
+            ManagerTheme::getLexicon('web_user_management_title'),
             '',
             'edit_user',
             'main',
@@ -672,13 +673,13 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuRoleManagement()
     {
-        if ($this->managerTheme->getCore()->hasPermission('edit_role')) {
+        if (ManagerTheme::getCore()->hasPermission('edit_role')) {
             $this->sitemenu['role_management_title'] = [
                 'role_management_title',
                 'users',
-                '<i class="' . $this->managerTheme->getStyle('icon_role') . '"></i>' . $this->managerTheme->getLexicon('role_management_title') . '<i class="' . $this->managerTheme->getStyle('icon_angle_right') . ' toggle"></i>',
+                '<i class="' . ManagerTheme::getStyle('icon_role') . '"></i>' . ManagerTheme::getLexicon('role_management_title') . '<i class="' . ManagerTheme::getStyle('icon_angle_right') . ' toggle"></i>',
                 'index.php?a=86',
-                $this->managerTheme->getLexicon('role_management_title'),
+                ManagerTheme::getLexicon('role_management_title'),
                 '',
                 'new_role,edit_role,delete_role',
                 'main',
@@ -693,16 +694,16 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuWebPermissions()
     {
-        if (!$this->managerTheme->getCore()->hasPermission('manage_groups')) {
+        if (!ManagerTheme::getCore()->hasPermission('manage_groups')) {
             return $this;
         }
 
         $this->sitemenu['web_permissions'] = [
             'web_permissions',
             'users',
-            '<i class="' . $this->managerTheme->getStyle('icon_web_user_access') . '"></i>' . $this->managerTheme->getLexicon('web_permissions'),
+            '<i class="' . ManagerTheme::getStyle('icon_web_user_access') . '"></i>' . ManagerTheme::getLexicon('web_permissions'),
             'index.php?a=91',
-            $this->managerTheme->getLexicon('web_permissions'),
+            ManagerTheme::getLexicon('web_permissions'),
             '',
             'web_access_permissions',
             'main',
@@ -719,9 +720,9 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['refresh_site'] = [
             'refresh_site',
             'tools',
-            '<i class="' . $this->managerTheme->getStyle('icon_recycle') . '"></i>' . $this->managerTheme->getLexicon('refresh_site'),
+            '<i class="' . ManagerTheme::getStyle('icon_recycle') . '"></i>' . ManagerTheme::getLexicon('refresh_site'),
             'index.php?a=26',
-            $this->managerTheme->getLexicon('refresh_site'),
+            ManagerTheme::getLexicon('refresh_site'),
             '',
             '',
             'main',
@@ -736,11 +737,11 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
                     // href
                     'btn btn-secondary',
                     // class or btn-success
-                    "modx.popup({url:'index.php?a=26', title:'" . $this->managerTheme->getLexicon('refresh_site') . "', icon: 'fa-recycle', iframe: 'ajax', selector: '.tab-page>.container', position: 'right top', width: 'auto', maxheight: '50%%', wrap: 'body' })",
+                    "modx.popup({url:'index.php?a=26', title:'" . ManagerTheme::getLexicon('refresh_site') . "', icon: 'fa-recycle', iframe: 'ajax', selector: '.tab-page>.container', position: 'right top', width: 'auto', maxheight: '50%%', wrap: 'body' })",
                     // onclick
-                    $this->managerTheme->getLexicon('refresh_site'),
+                    ManagerTheme::getLexicon('refresh_site'),
                     // title
-                    '<i class="' . $this->managerTheme->getStyle('icon_recycle') . '"></i>',
+                    '<i class="' . ManagerTheme::getStyle('icon_recycle') . '"></i>',
                     // innerHTML
                 ],
             ],
@@ -754,9 +755,9 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['search'] = [
             'search',
             'tools',
-            '<i class="' . $this->managerTheme->getStyle('icon_search') . '"></i>' . $this->managerTheme->getLexicon('search'),
+            '<i class="' . ManagerTheme::getStyle('icon_search') . '"></i>' . ManagerTheme::getLexicon('search'),
             'index.php?a=71',
-            $this->managerTheme->getLexicon('search'),
+            ManagerTheme::getLexicon('search'),
             '',
             '',
             'main',
@@ -770,16 +771,16 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuBkManager()
     {
-        if (!$this->managerTheme->getCore()->hasPermission('bk_manager')) {
+        if (!ManagerTheme::getCore()->hasPermission('bk_manager')) {
             return $this;
         }
 
         $this->sitemenu['bk_manager'] = [
             'bk_manager',
             'tools',
-            '<i class="' . $this->managerTheme->getStyle('icon_database') . '"></i>' . $this->managerTheme->getLexicon('bk_manager'),
+            '<i class="' . ManagerTheme::getStyle('icon_database') . '"></i>' . ManagerTheme::getLexicon('bk_manager'),
             'index.php?a=93',
-            $this->managerTheme->getLexicon('bk_manager'),
+            ManagerTheme::getLexicon('bk_manager'),
             '',
             'bk_manager',
             'main',
@@ -793,16 +794,16 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
 
     protected function menuRemoveLocks()
     {
-        if (!$this->managerTheme->getCore()->hasPermission('remove_locks')) {
+        if (!ManagerTheme::getCore()->hasPermission('remove_locks')) {
             return $this;
         }
 
         $this->sitemenu['remove_locks'] = [
             'remove_locks',
             'tools',
-            '<i class="' . $this->managerTheme->getStyle('icon_hourglass') . '"></i>' . $this->managerTheme->getLexicon('remove_locks'),
+            '<i class="' . ManagerTheme::getStyle('icon_hourglass') . '"></i>' . ManagerTheme::getLexicon('remove_locks'),
             'javascript:modx.removeLocks();',
-            $this->managerTheme->getLexicon('remove_locks'),
+            ManagerTheme::getLexicon('remove_locks'),
             '',
             'remove_locks',
             '',
@@ -819,9 +820,9 @@ class Frame extends AbstractController implements ManagerTheme\PageControllerInt
         $this->sitemenu['update_tree'] = [
             'update_tree',
             'tools',
-            '<i class="' . $this->managerTheme->getStyle('icon_sitemap') . '"></i>' . $this->managerTheme->getLexicon('update_tree'),
+            '<i class="' . ManagerTheme::getStyle('icon_sitemap') . '"></i>' . ManagerTheme::getLexicon('update_tree'),
             'index.php?a=95',
-            $this->managerTheme->getLexicon('update_tree'),
+            ManagerTheme::getLexicon('update_tree'),
             '',
             'update_tree',
             'main',
