@@ -4,26 +4,27 @@ declare(strict_types=1);
 
 namespace EvolutionCMS\Controllers;
 
-use EvolutionCMS\Interfaces\ManagerTheme;
+use EvolutionCMS\Facades\ManagerTheme;
+use EvolutionCMS\Interfaces\ManagerTheme\PageControllerInterface;
 use EvolutionCMS\Models\Category;
 use EvolutionCMS\Models\SiteModule;
 use EvolutionCMS\Support\ContextMenu;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\Collection;
 
-class Modules extends AbstractController implements ManagerTheme\PageControllerInterface
+class Modules extends AbstractController implements PageControllerInterface
 {
     /**
      * @var string
      */
-    protected $view = 'page.modules';
+    protected string $view = 'page.modules';
 
     /**
      * @return bool
      */
     public function canView(): bool
     {
-        return $this->managerTheme->getCore()
+        return ManagerTheme::getCore()
             ->hasAnyPermissions([
                 'exec_module',
                 'new_module',
@@ -54,9 +55,9 @@ class Modules extends AbstractController implements ManagerTheme\PageControllerI
         // context menu
         $cm = new ContextMenu('cntxm', 150);
 
-        $cm->addItem(__('global.run_module'), "js:menuAction(1)", $this->managerTheme->getStyle('icon_play'), (!$this->managerTheme->getCore()
+        $cm->addItem(__('global.run_module'), "js:menuAction(1)", ManagerTheme::getStyle('icon_play'), (!ManagerTheme::getCore()
             ->hasPermission('exec_module') ? 1 : 0));
-        if ($this->managerTheme->getCore()
+        if (ManagerTheme::getCore()
             ->hasAnyPermissions([
                 'new_module',
                 'edit_module',
@@ -64,11 +65,11 @@ class Modules extends AbstractController implements ManagerTheme\PageControllerI
             ])) {
             $cm->addSeparator();
         }
-        $cm->addItem(__('global.edit'), 'js:menuAction(2)', $this->managerTheme->getStyle('icon_edit'), (!$this->managerTheme->getCore()
+        $cm->addItem(__('global.edit'), 'js:menuAction(2)', ManagerTheme::getStyle('icon_edit'), (!ManagerTheme::getCore()
             ->hasPermission('edit_module') ? 1 : 0));
-        $cm->addItem(__('global.duplicate'), 'js:menuAction(3)', $this->managerTheme->getStyle('icon_clone'), (!$this->managerTheme->getCore()
+        $cm->addItem(__('global.duplicate'), 'js:menuAction(3)', ManagerTheme::getStyle('icon_clone'), (!ManagerTheme::getCore()
             ->hasPermission('new_module') ? 1 : 0));
-        $cm->addItem(__('global.delete'), 'js:menuAction(4)', $this->managerTheme->getStyle('icon_trash'), (!$this->managerTheme->getCore()
+        $cm->addItem(__('global.delete'), 'js:menuAction(4)', ManagerTheme::getStyle('icon_trash'), (!ManagerTheme::getCore()
             ->hasPermission('delete_module') ? 1 : 0));
 
         return [
