@@ -1,14 +1,12 @@
 <?php namespace EvolutionCMS\Controllers;
 
-use EvolutionCMS\Interfaces\ManagerTheme;
-use EvolutionCMS\Interfaces\ManagerThemeInterface;
-use EvolutionCMS\Models;
+use EvolutionCMS\Interfaces\ManagerTheme\PageControllerInterface;
 use EvolutionCMS\Models\ClosureTable;
 use EvolutionCMS\Models\SiteContent;
 
-class UpdateTree extends AbstractController implements ManagerTheme\PageControllerInterface
+class UpdateTree extends AbstractController implements PageControllerInterface
 {
-    protected $view = 'page.update_tree';
+    protected string $view = 'page.update_tree';
 
     /**
      * @var \EvolutionCMS\Interfaces\DatabaseInterface
@@ -23,18 +21,18 @@ class UpdateTree extends AbstractController implements ManagerTheme\PageControll
         return true;
     }
 
-    public function process() : bool
+    public function process(): bool
     {
-        $this->parameters['count'] = Models\SiteContent::query()->count();
+        $this->parameters['count'] = SiteContent::query()->count();
 
         $this->parameters['finish'] = 0;
-        if(isset($_POST['start'])){
+        if (isset($_POST['start'])) {
             $start = microtime(true);
-            $result = SiteContent::query()->where('parent',0)->get();
+            $result = SiteContent::query()->where('parent', 0)->get();
             ClosureTable::query()->truncate();
-            while($result->count() > 0){
+            while ($result->count() > 0) {
                 $parents = [];
-                foreach ($result as $item){
+                foreach ($result as $item) {
                     $descendant = $item->getKey();
                     $ancestor = isset($item->parent) ? $item->parent : $descendant;
                     $item->closure = new ClosureTable();
@@ -49,8 +47,6 @@ class UpdateTree extends AbstractController implements ManagerTheme\PageControll
             $this->parameters['finish'] = 1;
 
         }
-
-
 
         return true;
     }
