@@ -1,22 +1,22 @@
 <?php namespace EvolutionCMS\Controllers\Users;
 
 use EvolutionCMS\Controllers\AbstractController;
-use EvolutionCMS\Exceptions\ServiceActionException;
 use EvolutionCMS\Exceptions\ServiceValidationException;
-use EvolutionCMS\Models;
-use EvolutionCMS\Interfaces\ManagerTheme;
+use EvolutionCMS\Interfaces\ManagerTheme\PageControllerInterface;
+use EvolutionCMS\Models\UserAttribute;
 use Illuminate\Support\Facades\Lang;
+use EvolutionCMS\UserManager\Facades\UserManager;
 
-class DeleteUser extends AbstractController implements ManagerTheme\PageControllerInterface
+class DeleteUser extends AbstractController implements PageControllerInterface
 {
-    protected $view = '';
+    protected string $view = '';
 
     /**
      * {@inheritdoc}
      */
     public function canView(): bool
     {
-        return $this->managerTheme->getCore()->hasPermission('delete_user');
+        return ManagerTheme::getCore()->hasPermission('delete_user');
     }
 
     public function process() : bool
@@ -24,15 +24,15 @@ class DeleteUser extends AbstractController implements ManagerTheme\PageControll
         if($_GET['id'] == evo()->getLoginUserID()){
             EvolutionCMS()->webAlertAndQuit(Lang::get('global.delete_yourself'));
         }
-        $user = Models\UserAttribute::query()->where('internalKey', $_GET['id'])->first();
+        $user = UserAttribute::query()->where('internalKey', $_GET['id'])->first();
         if($user->role == 1){
-            $otherAdmin = Models\UserAttribute::query()->where('role', 1)->where('internalKey', '!=', $_GET['id'])->count();
+            $otherAdmin = UserAttribute::query()->where('role', 1)->where('internalKey', '!=', $_GET['id'])->count();
             if($otherAdmin == 0){
                 EvolutionCMS()->webAlertAndQuit(Lang::get('global.delete_last_admin'));
             }
         }
         try {
-            \UserManager::delete($_GET);
+            UserManager::delete($_GET);
         }catch (ServiceValidationException $exception){
             foreach ($exception->getValidationErrors() as $errors){
                 if(is_array($errors)){
