@@ -1,7 +1,7 @@
 <?php
 return [
-    'version' => '3.1.17', // Current version number
-    'release_date' => 'April 23, 2023', // Date of release
+    'version' => '3.2.0', // Current version number
+    'release_date' => 'May 7, 2023', // Date of release
     'branch' => 'Evolution CE', // Codebase name
-    'full_appname' => 'Evolution CE 3.1.17 (April 23, 2023)'
+    'full_appname' => 'Evolution CE 3.2.0 (May 7, 2023)'
 ];
