@@ -42,7 +42,7 @@ $errors = 0;
 
 
 // check PHP version
-define('PHP_MIN_VERSION', '7.4.0');
+define('PHP_MIN_VERSION', '8.1.0');
 $phpMinVersion = PHP_MIN_VERSION; // Maybe not necessary. For backward compatibility
 echo '<p>' . $_lang['checking_php_version'];
 // -1 if left is less, 0 if equal, +1 if left is higher
